pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "isagellm-benchmark"
version = "0.1.0.2"
description = "Benchmark Suite & E2E Testing for sageLLM"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "Private" }
authors = [{ name = "IntelliStream Team", email = "shuhao_zhang@hust.edu.cn" }]
keywords = ["llm", "inference", "benchmark", "testing", "domestic-hardware"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    # Core dependencies (sagellm-protocol is the base)
    "isagellm-protocol>=0.1.0",
    # CLI dependencies
    "click>=8.0.0",
    "rich>=13.0.0",
    # Async support
    "asyncio>=3.4.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=0.21.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0",
]
# Optional client dependencies
openai-client = ["openai>=1.0.0"]
vllm-client = ["vllm>=0.2.0", "openai>=1.0.0"]
lmdeploy-client = ["lmdeploy>=0.2.0", "httpx>=0.24.0"]
sagellm-client = ["isagellm-backend>=0.1.0"]
# All clients
all-clients = [
    "openai>=1.0.0",
    "vllm>=0.2.0",
    "lmdeploy>=0.2.0",
    "httpx>=0.24.0",
    "isagellm-backend>=0.1.0",
]

[project.scripts]
sagellm-benchmark = "sagellm_benchmark.cli:main"

[project.urls]
Homepage = "https://github.com/intellistream/sagellm-benchmark"
Repository = "https://github.com/intellistream/sagellm-benchmark"
Issues = "https://github.com/intellistream/sagellm-benchmark/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "e2e: End-to-end tests",
    "year1: Year 1 demo tests",
    "year2: Year 2 demo tests",
    "year3: Year 3 demo tests",
]
